Update to showroom and third eye sit functions
Showroom now spawns local tablets when in range of the pdm for everyone clientside instead of trying to do it server side( no need because the object doesnt move). Also, we now freeze the chair we are sitting in then unfreeze it once the animation is complete in order to keep it from sliding out from under you.